修复打包的一个报错

+++ Unity/Assets/Scripts/Core/Helper/FileHelper.cs
@@ -23,6 +23,10 @@ namespace ET
 		
 		public static void CleanDirectory(string dir)
 		{
+			if (!Directory.Exists(dir))
+			{
+				return;
+			}
 			foreach (string subdir in Directory.GetDirectories(dir))
 			{
 				Directory.Delete(subdir, true);
